Atualização da pasta Admin

+++ src/view/admin/menuAdm.php
@@ -1,31 +1,35 @@
-<!-- Menu nivel 2 -->
-<nav class="navDesk">
-    <ul>
-        <li><a href="#">RELATÓRIOS</a>
+<?php if ($id_perfil <= 3){ ?>
+    <!-- Menu nivel 2 -->
+    <nav class="navDesk">
+        <ul>
+            <li><a href="#">RELATÓRIOS</a>
+                <ul>
+                    <li><a href="#">Cadastro de Leitores</a></li>
+                    <li><a href="#">Cadastro de Sebos</a></li>
+                    <li><a href="#">Geral de Acervos</a></li>
+                    <li><a href="#">Duvidas</a></li>
+                </ul>
+            </li>
+            <li><a href="tratamento.php">TRATAMENTO DE DUVIDAS</a></li>
+            <li><a href="eventosComentarios.php">EVENTOS E COMENTARIOS</a></li>
+            <li><a href="controleUsuario.php">CONTROLE DE USUARIOS</a></li>
+        </ul>
+    <?php if ($id_perfil <= 2){ ?>
+        <!-- Menu nivel 1 -->
             <ul>
-                <li><a href="#">Cadastro de Leitores</a></li>
-                <li><a href="#">Cadastro de Sebos</a></li>
-                <li><a href="#">Geral de Acervos</a></li>
-                <li><a href="#">Duvidas</a></li>
+                <li><a href="apagar.php">APAGAR POSTAGENS INDEVIDAS</a></li>
+                <li><a href="cadastraEvento.php">CADASTRAR EVENTOS E NOTÍCIAS:</a></li>
+                <li><a href="Bloquear.html">BLOQUEAR USUÁRIOS</a></li>
+                <li><a href="aprovarSebo.php">APROVAR SEBOS PARCEIROS</a></li>
+                <li><a href="ExcluirSebo.html">EXCLUIR SEBOS PARCEIROS</a></li>
             </ul>
-        </li>
-        <li><a href="tratamento.php">TRATAMENTO DE DUVIDAS</a></li>
-        <li><a href="eventosComentarios.php">EVENTOS E COMENTARIOS</a></li>
-        <li><a href="controleUsuario.php">CONTROLE DE USUARIOS</a></li>
-    </ul>
-
-<!-- Menu nivel 1 -->
-    <ul>
-        <li><a href="apagar.php">APAGAR POSTAGENS INDEVIDAS</a></li>
-        <li><a href="cadastraEvento.php">CADASTRAR EVENTOS E NOTÍCIAS:</a></li>
-        <li><a href="Bloquear.html">BLOQUEAR USUÁRIOS</a></li>
-        <li><a href="aprovarSebo.php">APROVAR SEBOS PARCEIROS</a></li>
-        <li><a href="ExcluirSebo.html">EXCLUIR SEBOS PARCEIROS</a></li>
-    </ul>
-
-<!-- Menu master -->
-    <ul>
-        <li><a href="aprovarISBN.php">APROVAR CADASTROS SEM ISBN</a></li>
-        <li><a href="banir.php">BANIR USUARIOS</a></li>
-    </ul>
-</nav>
+        <?php if ($id_perfil <= 1){ ?>
+            <!-- Menu master -->
+                <ul>
+                    <li><a href="aprovarISBN.php">APROVAR CADASTROS SEM ISBN</a></li>
+                    <li><a href="banir.php">BANIR USUARIOS</a></li>
+                </ul>
+            </nav>
+    <?php }
+    }
+}?>
